Most of intended functionality has been added. Project is mostly finsihed.

+++ src/pages/Journal.jsx
@@ -1,11 +1,11 @@
 import { useState, useEffect } from 'react';
 import '../css/journal.css';
 import WorkoutWrapper from '../components/WorkoutWrapper';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
 const Journal = () => {
   const [workouts, setWorkouts] = useState([]);
-  const { id } = useParams();
+  const [search, setSearch] = useState('');
 
   const navigate = useNavigate();
 
@@ -26,6 +26,10 @@ const Journal = () => {
 
   // delete a workout
   const deleteWorkout = async (id) => {
+    const confirmDelete = window.confirm('Press OK if you want to delete.');
+
+    if (!confirmDelete) return;
+
     const res = await fetch(`http://localhost:5000/workouts/${id}`, {
       method: 'DELETE',
     });
@@ -36,8 +40,12 @@ const Journal = () => {
   // edit a workout
   const editWorkout = async (id) => {
     navigate(`/editworkout/${id}`);
-    console.log('edited', id);
   };
+
+  // filter through workouts by name
+  const filterWorkouts = workouts.filter((workout) =>
+    JSON.stringify(workout).toLowerCase().includes(search.toLowerCase())
+  );
 
   return (
     <div className='journal-container'>
@@ -45,44 +53,40 @@ const Journal = () => {
         <h1>Welcome to your workout journal!</h1>
         <h3>All workouts your have created will show here.</h3>
       </div>
+      <div className='search-container'>
+        <input
+          type='text'
+          className='search-input'
+          value={search}
+          placeholder='Search'
+          onChange={(e) => setSearch(e.target.value)}
+        />
+      </div>
       <div className='journal-workouts'>
-        {workouts.map((workout) => (
-          <ul key={workout.id}>
-            <WorkoutWrapper>
-              <h3
-                style={{
-                  backgroundColor: 'black',
-                  borderRadius: '8px',
-                  color: 'whitesmoke',
-                }}
-              >
-                Workout Name: <li>{workout.name}</li>
-              </h3>
-
-              <h3>Exercises: </h3>
-              <li>{workout.exercise}</li>
-              <h3>Date:</h3>
-              <li>
-                {workout.day}/{workout.month}/{workout.year}
-              </li>
-              <button
-                onClick={() => {
-                  deleteWorkout(workout.id);
-                }}
-              >
-                Delete Workout
-              </button>
-              <button
-                type='button'
-                id='edit-button'
-                onClick={() => {
-                  editWorkout(workout.id);
-                }}
-              >
-                Edit
-              </button>
-            </WorkoutWrapper>
-          </ul>
+        {filterWorkouts.map((workout) => (
+          <WorkoutWrapper key={workout.id}>
+            <h2>Workout Name: {workout.name}</h2>
+            <h3>Exercises: </h3>
+            <p>{workout.exercises}</p>
+            <h3>Date:</h3>
+            <p>{workout.date}</p>
+            <button
+              onClick={() => {
+                deleteWorkout(workout.id);
+              }}
+            >
+              Delete Workout
+            </button>
+            <button
+              type='button'
+              id='edit-button'
+              onClick={() => {
+                editWorkout(workout.id);
+              }}
+            >
+              Edit
+            </button>
+          </WorkoutWrapper>
         ))}
       </div>
     </div>
